# src/main/java/com/springtutorials/dto/UserDTO.java
package com.springtutorials.dto;



import jakarta.validation.constraints.Email;

public class UserDTO {
	
	private Long userid;
	private String username;
	@Email
	private String email;
	private String password;
	private String phone;
	
	
    public String getPhone() {
		return phone;
	}
	public void setPhone(String phone) {
		this.phone = phone;
	}
	
	public String getUsername() {
		return username;
	}
	public void setUsername(String username) {
		this.username = username;
	}
	public Long getUserid() {
		return userid;
	}
	public void setUserid(Long userid) {
		this.userid = userid;
	}
	public String getEmail() {
		return email;
	}
	public void setEmail(String email) {
		this.email = email;
	}
	public String getPassword() {
		return password;
	}
	public void setPassword(String password) {
		this.password = password;
	}
	
	
	

}