# src/main/resources/static/swagger.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "OpenAPI definition",
		"version": "v0"
	},
	"servers": [
		{
			"url": "http://localhost:8080",
			"description": "Generated server url"
		}
	],
	"tags": [
		{
			"name": "e-commerce user Orders",
			"description": "Api for order of user"
		},
		{
			"name": "Authentication-for Roles",
			"description": "Authentication api for the roles"
		},
		{
			"name": "e-commerce user Cart",
			"description": "Api for cart of user"
		},
		{
			"name": "User e-commerce",
			"description": "API for e-commerce"
		},
		{
			"name": "e-commerce Product",
			"description": "Api for product of user"
		}
	],
	"paths": {
		"/api/users/update/{id}": {
			"put": {
				"tags": [
					"User e-commerce"
				],
				"summary": "update all users",
				"operationId": "updateUser",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/User"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				}
			}
		},
		"/api/users/me/update": {
			"put": {
				"tags": [
					"User e-commerce"
				],
				"summary": "update the user profile",
				"operationId": "updatemyprofile",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/User"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				}
			}
		},
		"/api/products/update": {
			"put": {
				"tags": [
					"e-commerce Product"
				],
				"summary": "update the product",
				"operationId": "updateProduct",
				"parameters": [],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Product"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/Product"
								}
							}
						}
					}
				}
			}
		},
		"/api/orders/update": {
			"put": {
				"tags": [
					"e-commerce user Orders"
				],
				"summary": "update the order",
				"operationId": "updateOrder",
				"parameters": [],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Order"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/Order"
								}
							}
						}
					}
				}
			}
		},
		"/api/users/register": {
			"post": {
				"tags": [
					"User e-commerce"
				],
				"summary": "user registration",
				"operationId": "registerUser",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserDTO"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				}
			}
		},
		"/api/products/multiple": {
			"post": {
				"tags": [
					"e-commerce Product"
				],
				"summary": "create multiple products",
				"operationId": "createMultipleProducts",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "array",
								"items": {
									"$ref": "#/components/schemas/Product"
								}
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Product"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/products/create": {
			"post": {
				"tags": [
					"e-commerce Product"
				],
				"summary": "create product",
				"operationId": "createProduct",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Product"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/Product"
								}
							}
						}
					}
				}
			}
		},
		"/api/orders/create": {
			"post": {
				"tags": [
					"e-commerce user Orders"
				],
				"summary": "create the order",
				"operationId": "createOrder",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Order"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/Order"
								}
							}
						}
					}
				}
			}
		},
		"/api/auth/login": {
			"post": {
				"tags": [
					"Authentication-for Roles"
				],
				"summary": "login api",
				"operationId": "login",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/User"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"type": "object"
								}
							}
						}
					}
				}
			}
		},
		"/api/users": {
			"get": {
				"tags": [
					"User e-commerce"
				],
				"summary": "get all users",
				"operationId": "getAllUsers",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/User"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/users/{id}": {
			"get": {
				"tags": [
					"User e-commerce"
				],
				"summary": "get users by id",
				"operationId": "getUserById",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				}
			}
		},
		"/api/users/me": {
			"get": {
				"tags": [
					"User e-commerce"
				],
				"summary": "get the user profile ",
				"operationId": "getMyProfile",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				}
			}
		},
		"/api/users/email/{email}": {
			"get": {
				"tags": [
					"User e-commerce"
				],
				"summary": "get users by email",
				"operationId": "getUserByEmail",
				"parameters": [
					{
						"name": "email",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				}
			}
		},
		"/api/products": {
			"get": {
				"tags": [
					"e-commerce Product"
				],
				"summary": "get all products",
				"operationId": "getAllProducts",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Product"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/products/{id}": {
			"get": {
				"tags": [
					"e-commerce Product"
				],
				"summary": "get product by id",
				"operationId": "getProductById",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/Product"
								}
							}
						}
					}
				}
			}
		},
		"/api/products/stock": {
			"get": {
				"tags": [
					"e-commerce Product"
				],
				"summary": "get the product stock",
				"operationId": "productByStock",
				"parameters": [
					{
						"name": "stock",
						"in": "query",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Product"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/products/price-range": {
			"get": {
				"tags": [
					"e-commerce Product"
				],
				"summary": "get the product by the price range",
				"operationId": "getProductsByPriceRange",
				"parameters": [
					{
						"name": "min",
						"in": "query",
						"required": true,
						"schema": {
							"type": "number",
							"format": "double"
						}
					},
					{
						"name": "max",
						"in": "query",
						"required": true,
						"schema": {
							"type": "number",
							"format": "double"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Product"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/products/filter": {
			"get": {
				"tags": [
					"e-commerce Product"
				],
				"summary": "filtering the produts",
				"operationId": "getFilteredProducts",
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer",
							"format": "int32",
							"default": 0
						}
					},
					{
						"name": "size",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer",
							"format": "int32",
							"default": 10
						}
					},
					{
						"name": "sortBy",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"default": "id"
						}
					},
					{
						"name": "stock",
						"in": "query",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					},
					{
						"name": "min",
						"in": "query",
						"required": true,
						"schema": {
							"type": "number",
							"format": "double"
						}
					},
					{
						"name": "max",
						"in": "query",
						"required": true,
						"schema": {
							"type": "number",
							"format": "double"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/PageProduct"
								}
							}
						}
					}
				}
			}
		},
		"/api/orders": {
			"get": {
				"tags": [
					"e-commerce user Orders"
				],
				"summary": "get all orders",
				"operationId": "getAllOrders",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Order"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/orders/{id}": {
			"get": {
				"tags": [
					"e-commerce user Orders"
				],
				"summary": "get order by order id",
				"operationId": "getOrderById",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/Order"
								}
							}
						}
					}
				}
			}
		},
		"/api/orders/user/{userId}": {
			"get": {
				"tags": [
					"e-commerce user Orders"
				],
				"summary": "order by user id",
				"operationId": "getOrdersByUserId",
				"parameters": [
					{
						"name": "userId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Order"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/orders/status/{status}": {
			"get": {
				"tags": [
					"e-commerce user Orders"
				],
				"summary": "get Order status",
				"operationId": "getOrdersByStatus",
				"parameters": [
					{
						"name": "status",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Order"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/users/delete/{id}": {
			"delete": {
				"tags": [
					"User e-commerce"
				],
				"summary": "delete all users",
				"operationId": "deleteUser",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				}
			}
		},
		"/api/products/delete/{id}": {
			"delete": {
				"tags": [
					"e-commerce Product"
				],
				"summary": "delete the product",
				"operationId": "deleteProductbyId",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"type": "object"
								}
							}
						}
					}
				}
			}
		},
		"/api/orders/delete/{id}": {
			"delete": {
				"tags": [
					"e-commerce user Orders"
				],
				"summary": "delete order api",
				"operationId": "deleteOrder",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				}
			}
		},
		"/api/cart/delete/{id}": {
			"delete": {
				"tags": [
					"e-commerce user Cart"
				],
				"summary": "delete cart api",
				"operationId": "deleteCartProduct",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
			"User": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer",
						"format": "int64"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"role": {
						"type": "string",
						"enum": [
							"CUSTOMER",
							"ADMIN"
						]
					}
				}
			},
			"Product": {
				"type": "object",
				"properties": {
					"productid": {
						"type": "integer",
						"format": "int64"
					},
					"productname": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"stock": {
						"type": "integer",
						"format": "int32"
					}
				}
			},
			"Order": {
				"type": "object",
				"properties": {
					"orderid": {
						"type": "integer",
						"format": "int64"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"items": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/OrderItem"
						}
					},
					"orderDate": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"type": "string",
						"enum": [
							"PLACED",
							"SHIPPED",
							"DELIVERED",
							"CANCELLED"
						]
					},
					"totalAmount": {
						"type": "number",
						"format": "double"
					},
					"shippingAddress": {
						"$ref": "#/components/schemas/ShippingAddress"
					},
					"paymentMode": {
						"type": "string",
						"enum": [
							"COD",
							"UPI",
							"CARD",
							"NETBANKING"
						]
					}
				}
			},
			"OrderItem": {
				"type": "object",
				"properties": {
					"itemid": {
						"type": "integer",
						"format": "int64"
					},
					"product": {
						"$ref": "#/components/schemas/Product"
					},
					"quantity": {
						"type": "integer",
						"format": "int32"
					},
					"price": {
						"type": "number",
						"format": "double"
					}
				}
			},
			"ShippingAddress": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer",
						"format": "int64"
					},
					"state": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"street": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"zipcode": {
						"type": "string"
					}
				}
			},
			"UserDTO": {
				"type": "object",
				"properties": {
					"userid": {
						"type": "integer",
						"format": "int64"
					},
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					}
				}
			},
			"PageProduct": {
				"type": "object",
				"properties": {
					"totalElements": {
						"type": "integer",
						"format": "int64"
					},
					"totalPages": {
						"type": "integer",
						"format": "int32"
					},
					"size": {
						"type": "integer",
						"format": "int32"
					},
					"content": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Product"
						}
					},
					"number": {
						"type": "integer",
						"format": "int32"
					},
					"sort": {
						"$ref": "#/components/schemas/SortObject"
					},
					"first": {
						"type": "boolean"
					},
					"last": {
						"type": "boolean"
					},
					"numberOfElements": {
						"type": "integer",
						"format": "int32"
					},
					"pageable": {
						"$ref": "#/components/schemas/PageableObject"
					},
					"empty": {
						"type": "boolean"
					}
				}
			},
			"PageableObject": {
				"type": "object",
				"properties": {
					"offset": {
						"type": "integer",
						"format": "int64"
					},
					"sort": {
						"$ref": "#/components/schemas/SortObject"
					},
					"pageSize": {
						"type": "integer",
						"format": "int32"
					},
					"paged": {
						"type": "boolean"
					},
					"pageNumber": {
						"type": "integer",
						"format": "int32"
					},
					"unpaged": {
						"type": "boolean"
					}
				}
			},
			"SortObject": {
				"type": "object",
				"properties": {
					"empty": {
						"type": "boolean"
					},
					"sorted": {
						"type": "boolean"
					},
					"unsorted": {
						"type": "boolean"
					}
				}
			}
		}
	}
}